Refactor the feed code
