--- app/src/main/java/br/com/zup/hellozupper/ui/Constants.kt
@@ -1,8 +1,0 @@
-package br.com.zup.hellozupper.ui
-
-const val MESSAGE_SUCCESS_NEWS_READ_ADDED = "Notícia marcada como lida"
-const val MESSAGE_FAIL_NEWS_READ_ADDED = "Notícia não pode ser marcada como lida"
-const val MESSAGE_FAIL_LOAD_NEWS_LIST = "Erro ao carregar notícias"
-const val MESSAGE_SUCCESS_NEWS_READ_DELETED = "Notícia marcada como não lida"
-const val MESSAGE_FAIL_NEWS_READ_DELETED = "Notícia não pode ser marcada como não lida"
-const val MESSAGE_EMPTY_NEWS_LIST = "Ops! Não encontramos notícias com esses termos."
