+++ app/src/main/java/br/com/zup/hellozupper/utils/Contants.kt
@@ -16,5 +16,10 @@ const val PROGRAMAS = "Programas"
 const val ERRO_API_PILLARS = "Não foi possível carregar a lista de pilares"
 const val NOSSO_DNA = "Nosso DNA"
 const val KEY_PILLAR = "Chave pillar"
-
+const val MESSAGE_SUCCESS_NEWS_READ_ADDED = "Notícia marcada como lida"
+const val MESSAGE_FAIL_NEWS_READ_ADDED = "Notícia não pode ser marcada como lida"
+const val MESSAGE_FAIL_LOAD_NEWS_LIST = "Erro ao carregar notícias"
+const val MESSAGE_SUCCESS_NEWS_READ_DELETED = "Notícia marcada como não lida"
+const val MESSAGE_FAIL_NEWS_READ_DELETED = "Notícia não pode ser marcada como não lida"
+const val MESSAGE_EMPTY_NEWS_LIST = "Ops! Não encontramos notícias com esses termos."
